WIP: move backlog.go into core.go
