+++ consensus/tendermint/core/core.go
@@ -62,6 +62,16 @@ var (
 	errFailedDecodePrecommit = errors.New("failed to decode COMMIT")
 )
 
+var (
+	// msgPriority is defined for calculating processing priority to speedup consensus
+	// msgProposal > msgPrecommit > msgPrevote
+	msgPriority = map[uint64]int{
+		msgProposal:  1,
+		msgPrecommit: 2,
+		msgPrevote:   3,
+	}
+)
+
 // New creates an Istanbul consensus core
 func New(backend tendermint.Backend, config *tendermint.Config) Engine {
 	c := &core{
@@ -350,4 +360,139 @@ func timeoutPrevote(round int64) time.Duration {
 
 func timeoutPrecommit(round int64) time.Duration {
 	return initialProposeTimeout + time.Duration(round)*time.Second
+}
+
+//---------------------------------------Backlog---------------------------------------
+
+// checkMessage checks the message state
+// return errInvalidMessage if the message is invalid
+// return errFutureMessage if the message view is larger than current view
+// return errOldMessage if the message view is smaller than current view
+func (c *core) checkMessage(msgCode uint64, view *tendermint.View) error {
+	if view == nil || view.Sequence == nil || view.Round == nil {
+		return errInvalidMessage
+	}
+
+	if view.Cmp(c.currentView()) > 0 {
+		return errFutureMessage
+	}
+
+	if view.Cmp(c.currentView()) < 0 {
+		return errOldMessage
+	}
+
+	// StateAcceptRequest only accepts msgProposal
+	// other messages are future messages
+	if c.state == StateAcceptRequest {
+		if msgCode > msgProposal {
+			return errFutureMessage
+		}
+		return nil
+	}
+
+	// For states(StateProposeDone, StatePrevoteDone, StatePrecommitDone),
+	// can accept all message types if processing with same view
+	return nil
+}
+
+func (c *core) storeBacklog(msg *message, src tendermint.Validator) {
+	logger := c.logger.New("from", src, "state", c.state)
+
+	if src.Address() == c.Address() {
+		logger.Warn("Backlog from self")
+		return
+	}
+
+	logger.Trace("Store future message")
+
+	c.backlogsMu.Lock()
+	defer c.backlogsMu.Unlock()
+
+	backlog := c.backlogs[src]
+	if backlog == nil {
+		backlog = prque.New()
+	}
+	switch msg.Code {
+	case msgProposal:
+		var p *tendermint.Proposal
+		err := msg.Decode(&p)
+		if err == nil {
+			backlog.Push(msg, toPriority(msg.Code, p.View))
+		}
+		// for msgRoundChange, msgPrevote and msgPrecommit cases
+	default:
+		var p *tendermint.Subject
+		err := msg.Decode(&p)
+		if err == nil {
+			backlog.Push(msg, toPriority(msg.Code, p.View))
+		}
+	}
+	c.backlogs[src] = backlog
+}
+
+func (c *core) processBacklog() {
+	c.backlogsMu.Lock()
+	defer c.backlogsMu.Unlock()
+
+	for src, backlog := range c.backlogs {
+		if backlog == nil {
+			continue
+		}
+
+		logger := c.logger.New("from", src, "state", c.state)
+		var isFuture bool
+
+		// We stop processing if
+		//   1. backlog is empty
+		//   2. The first message in queue is a future message
+		for !(backlog.Empty() || isFuture) {
+			m, prio := backlog.Pop()
+			msg := m.(*message)
+			var view *tendermint.View
+			switch msg.Code {
+			case msgProposal:
+				var m *tendermint.Proposal
+				err := msg.Decode(&m)
+				if err == nil {
+					view = m.View
+				}
+				// for msgRoundChange, msgPrevote and msgPrecommit cases
+			default:
+				var sub *tendermint.Subject
+				err := msg.Decode(&sub)
+				if err == nil {
+					view = sub.View
+				}
+			}
+			if view == nil {
+				logger.Debug("Nil view", "msg", msg)
+				continue
+			}
+			// Push back if it's a future message
+			err := c.checkMessage(msg.Code, view)
+			if err != nil {
+				if err == errFutureMessage {
+					logger.Trace("Stop processing backlog", "msg", msg)
+					backlog.Push(msg, prio)
+					isFuture = true
+					break
+				}
+				logger.Trace("Skip the backlog event", "msg", msg, "err", err)
+				continue
+			}
+			logger.Trace("Post backlog event", "msg", msg)
+
+			go c.sendEvent(backlogEvent{
+				src: src,
+				msg: msg,
+			})
+		}
+	}
+}
+
+func toPriority(msgCode uint64, view *tendermint.View) float32 {
+	// FIXME: round will be reset as 0 while new sequence
+	// 10 * Round limits the range of message code is from 0 to 9
+	// 1000 * Sequence limits the range of round is from 0 to 99
+	return -float32(view.Sequence.Uint64()*1000 + view.Round.Uint64()*10 + uint64(msgPriority[msgCode]))
 }
